requirements.txt
torch
torchvision
numpy
opencv-python
pillow
pandas
scipy
tqdm
